ページネーションコンポーネントの共通化
Fixes #1005
